server/requirements_server.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.2.0
torch>=2.2.0
torchvision>=0.17.0
facenet-pytorch>=2.6.0
numpy>=1.26.0
pymongo>=4.6.0
python-dotenv>=1.0.0
httpx>=0.26.0
opencv-python>=4.9.0
scipy>=1.12.0
